add fragments to lesson and standalone template
- added utils function to addFragment blocks
